+++ aimiaobi-20230801/ChangeLog.md
@@ -1,3 +1,8 @@
+2025-08-04 Version: 1.32.1
+- Update API GetSmartAuditResult: add response parameters Body.Data.ErrorMessage.
+- Update API SubmitSmartAudit: add request parameters ImageUrlList.
+
+
 2025-08-01 Version: 1.32.0
 - Support API ConfirmAndPostProcessAuditNote.
 - Support API DeleteAuditNote.